refactor(Mypage): 세부기능별 폴더구조 개편
