+++ src/feature/mypage/mypage.main/components/typography/MyPageInfoItem.tsx
