+++ src/feature/mypage/mypage.main/components/typography/MyPageSectionTitle.tsx
